type VaryOptions.defaultVariant as a string
more of a toss up compared to previous commits but the idea is the same:
I think it'd be counterintuitive to have developers think they can pass
in a boolean and we don't pass a boolean in any calls here. Out in the
wild, I'd suspect that we'd have little use for literally having a
'true' boolean shorthand (i.e. just true) but :shrug: